refactor(vscode): split client options and node rename
