+++ editors/vscode/src/extension.ts
@@ -6,14 +6,11 @@ import { promisify } from 'node:util';
 import * as vscode from 'vscode';
 import {
   LanguageClient,
-  type LanguageClientOptions,
-  RevealOutputChannelOn,
   type ServerOptions,
 } from 'vscode-languageclient/node';
 
 import { registerDiagnosticActions } from './diagnosticActions';
 import { profileDiagnosticsCommand, registerProfilingCommand } from './profiling';
-import { serverInitializationOptions } from './initializationOptions';
 import {
   DEFAULT_PROJECT_CONFIG_TEXT,
   PROJECT_CONFIG_FILE_NAMES,
@@ -31,6 +28,8 @@ import {
   VideStatusController,
 } from './videStatus';
 import type { ServerStatus } from './status';
+import { createNodeClientOptions } from './common/clientOptions';
+import { createProvideExpandedRenameEdits } from './node/renameMiddleware';
 import {
   createServerEnv,
   readConfiguration,
@@ -49,9 +48,6 @@ const showServerVersionCommand = 'vide.showServerVersion';
 const showQiheOutputCommand = 'vide.showQiheOutput';
 const runQiheAnalysisCommand = 'vide.runQiheAnalysis';
 const runQiheAnalysisRequest = 'vide.server.runQiheAnalysis';
-const renameExpansionInfoRequest = 'vide.server.renameExpansionInfo';
-const expandedRenameRequest = 'vide.server.expandedRename';
-const renameConflictInfoRequest = 'vide.server.renameConflictInfo';
 const qiheStatusNotification = 'vide/qiheStatus';
 const qiheLogNotification = 'vide/qiheLog';
 const qiheAnalysisIcon = '$(beaker)';
@@ -293,145 +289,6 @@ function registerProjectStatusNotifications(languageClient: LanguageClient): voi
   });
 }
 
-function includeDeclarationInReferences(document: vscode.TextDocument): boolean {
-  return (
-    vscode.workspace
-      .getConfiguration('vide', document)
-      .get<boolean>('references.includeDeclaration') ?? true
-  );
-}
-
-type RenameExpansionInfo = {
-  additionalSymbols: number;
-};
-
-type RenameConflictInfo = {
-  conflicts: number;
-};
-
-function emptyRenameEdit(): vscode.WorkspaceEdit {
-  return new vscode.WorkspaceEdit();
-}
-
-async function confirmRenameCollision(
-  textDocumentPosition: unknown,
-  newName: string,
-  recursive: boolean,
-  token: vscode.CancellationToken,
-): Promise<boolean> {
-  const languageClient = client;
-  if (!languageClient) {
-    return true;
-  }
-
-  const info = await languageClient.sendRequest<RenameConflictInfo>(
-    'workspace/executeCommand',
-    {
-      command: renameConflictInfoRequest,
-      arguments: [{ textDocumentPosition, newName, recursive }],
-    },
-    token,
-  );
-
-  if (info.conflicts === 0) {
-    return true;
-  }
-
-  const continueAction = vscode.l10n.t('Continue Rename');
-  const cancelAction = vscode.l10n.t('Cancel');
-  const selected = await vscode.window.showWarningMessage(
-    vscode.l10n.t(
-      'Renaming to "{0}" may collide with {1} existing symbol(s).',
-      newName,
-      info.conflicts,
-    ),
-    continueAction,
-    cancelAction,
-  );
-  return selected === continueAction;
-}
-
-async function provideExpandedRenameEdits(
-  document: vscode.TextDocument,
-  position: vscode.Position,
-  newName: string,
-  token: vscode.CancellationToken,
-  next: (
-    document: vscode.TextDocument,
-    position: vscode.Position,
-    newName: string,
-    token: vscode.CancellationToken,
-  ) => vscode.ProviderResult<vscode.WorkspaceEdit>,
-): Promise<vscode.WorkspaceEdit | null | undefined> {
-  const languageClient = client;
-  if (!languageClient) {
-    return await next(document, position, newName, token);
-  }
-
-  const textDocumentPosition = {
-    textDocument: languageClient.code2ProtocolConverter.asTextDocumentIdentifier(document),
-    position: languageClient.code2ProtocolConverter.asPosition(position),
-  };
-  const standardRename = async (): Promise<vscode.WorkspaceEdit | null | undefined> => {
-    if (!(await confirmRenameCollision(textDocumentPosition, newName, false, token))) {
-      return emptyRenameEdit();
-    }
-    return await next(document, position, newName, token);
-  };
-
-  let info: RenameExpansionInfo | undefined;
-  try {
-    info = await languageClient.sendRequest<RenameExpansionInfo>(
-      'workspace/executeCommand',
-      {
-        command: renameExpansionInfoRequest,
-        arguments: [{ textDocumentPosition }],
-      },
-      token,
-    );
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error);
-    log(`[WARN] Falling back to standard rename: ${message}`);
-  }
-
-  if (!info || info.additionalSymbols === 0) {
-    return await standardRename();
-  }
-
-  const recursiveAction = vscode.l10n.t('Rename Connected Ports/Signals');
-  const localAction = vscode.l10n.t('Only This Symbol');
-  const selected = await vscode.window.showInformationMessage(
-    vscode.l10n.t(
-      'Rename {0} connected port/signal symbol(s) as well?',
-      info.additionalSymbols,
-    ),
-    recursiveAction,
-    localAction,
-  );
-
-  if (selected === localAction) {
-    return await standardRename();
-  }
-
-  if (selected !== recursiveAction) {
-    return emptyRenameEdit();
-  }
-
-  if (!(await confirmRenameCollision(textDocumentPosition, newName, true, token))) {
-    return emptyRenameEdit();
-  }
-
-  const edit = await languageClient.sendRequest(
-    'workspace/executeCommand',
-    {
-      command: expandedRenameRequest,
-      arguments: [{ textDocumentPosition, newName }],
-    },
-    token,
-  );
-  return await languageClient.protocol2CodeConverter.asWorkspaceEdit(edit as never, token);
-}
-
 type ProjectConfigTarget = {
   folderName: string;
   configPath: string;
@@ -625,27 +482,11 @@ async function createClient(context: vscode.ExtensionContext): Promise<LanguageC
     },
   };
 
-  const clientOptions: LanguageClientOptions = {
-    documentSelector: [
-      { scheme: 'file', language: 'verilog' },
-      { scheme: 'file', language: 'systemverilog' },
-    ],
-    synchronize: {
-      configurationSection: ['vide'],
-    },
+  const clientOptions = createNodeClientOptions({
     outputChannel: channel,
-    traceOutputChannel: channel,
-    revealOutputChannelOn: RevealOutputChannelOn.Never,
-    initializationOptions: serverInitializationOptions(vscode.workspace.getConfiguration('vide')),
-    middleware: {
-      provideReferences: async (document, position, options, token, next) => {
-        options.includeDeclaration = includeDeclarationInReferences(document);
-        return await next(document, position, options, token);
-      },
-      provideRenameEdits: provideExpandedRenameEdits,
-    },
-    ...(config.trace !== 'off' && { trace: config.trace }),
-  };
+    trace: config.trace,
+    provideRenameEdits: createProvideExpandedRenameEdits(() => client, log),
+  });
 
   log('[INFO] Creating LanguageClient instance...');
   return new LanguageClient(
